+++ app/src/components/features/rules/TestThisRule/index.ts
@@ -1,3 +1,2 @@
 export { TestThisRule } from "./TestThisRule";
 export * from "./utils/testReports";
-export * from "./utils/widget";
