+++ app/src/features/settings/components/GlobalSettings/components/ImplicitRuleTesting/types.ts
@@ -1,0 +1,4 @@
+export type ImplicitRuleTestingWidgetConfig = {
+  visibility: string;
+  ruleTypes: string[];
+};
