+++ common/constants.js
@@ -28,8 +28,6 @@ CONSTANTS.PUBLIC_NAMESPACE = "__REQUESTLY__";
 
 CONSTANTS.CONSOLE_LOGGER_ENABLED = "console-logger-enabled";
 
-CONSTANTS.IMPLICIT_RULE_TESTING_WIDGET_CONFIG = "implicit_rule_testing_widget_config";
-
 CONSTANTS.IMPLICIT_RULE_TESTING_WIDGET_VISIBILITY = {
   ALL: "all",
   SPECIFIC: "specific",
@@ -213,6 +211,7 @@ CONSTANTS.STORAGE_KEYS = {
   USE_EVENTS_ENGINE: "useEventsEngine",
   SEND_EXECUTION_EVENTS: "sendExecutionEvents",
   TEST_REPORTS: "testReports",
+  IMPLICIT_RULE_TESTING_WIDGET_CONFIG: "implicit_rule_testing_widget_config",
 };
 
 CONSTANTS.REQUEST_STATE = {
